Split off classes into files
