+++ zypp-plugin/snapper_zypp_plugin.cc
@@ -6,8 +6,6 @@
 
 #include <boost/format.hpp>
 using boost::format;
-#include <boost/regex.hpp>
-#include <boost/algorithm/string/trim.hpp>
 
 #include <iostream>
 #include <map>
@@ -18,6 +16,8 @@ using namespace std;
 #include "snapper/Exception.h"
 using snapper::Exception;
 using snapper::CodeLocation;
+
+#include "zypp_commit_plugin.h"
 
 // this is copied from command.cc, not part of any library now
 
@@ -98,125 +98,6 @@ public:
 };
 Logging logging;
 
-// Plugin message aka frame
-// https://doc.opensuse.org/projects/libzypp/SLE12SP2/zypp-plugins.html
-struct Message {
-    string command;
-    map<string, string> headers;
-    string body;
-};
-
-class ZyppPlugin {
-public:
-    virtual int main();
-    virtual Message dispatch(const Message&);
-    void answer(const Message&);
-    virtual ~ZyppPlugin() {}
-};
-
-void ZyppPlugin::answer(const Message& msg) {
-    cout << msg.command << endl;
-    for(auto it: msg.headers) {
-	cout << it.first << ':' << it.second << endl;
-    }
-    cout << endl;
-    cout << msg.body << '\0';
-    cout.flush();
-}
-
-int ZyppPlugin::main() {
-    enum class State {
-		      Start,
-		      Headers,
-		      Body
-    } state = State::Start;
-
-    Message msg;
-    static const boost::regex rx_word("[A-Za-z0-9_]+");
-    while(!cin.eof()) {
-	string line;
-
-	getline(cin, line);
-	boost::trim_right(line);
-
-	if (state == State::Start) {
-	    if (line.empty())
-		continue;
-	    if (boost::regex_match(line, rx_word)) {
-		msg = Message();
-		msg.command = line;
-		state = State::Headers;
-	    }
-	    else {
-		throw "FIXME: expected a command";
-	    }
-	}
-	else if (state == State::Headers) {
-	    if (line.empty()) {
-		getline(cin, msg.body, '\0');
-
-		Message reply = dispatch(msg);
-		answer(reply);
-
-		state = State::Start;
-	    }
-	    else {
-		//string key, value;
-		// TODO: parse the header
-		//msg.headers[key] = value;
-	    }
-	}
-    }
-    return 0;
-}
-
-Message ZyppPlugin::dispatch(const Message& msg) {
-    Message a;
-    a.command = "_ENOMETHOD";
-    a.headers["Command"] = msg.command;
-    return a;
-}
-
-class ZyppCommitPlugin : public ZyppPlugin {
-public:
-    Message dispatch(const Message& msg) override;
-
-    virtual Message plugin_begin(const Message& m) {
-	return ack();
-    }
-    virtual Message plugin_end(const Message& m) {
-	return ack();
-    }
-    virtual Message commit_begin(const Message& m) {
-	return ack();
-    }
-    virtual Message commit_end(const Message& m) {
-	return ack();
-    }
-
-    Message ack() {
-	Message a;
-	a.command = "ACK";
-	return a;
-    }
-};
-
-Message ZyppCommitPlugin::dispatch(const Message& msg) {
-    if (msg.command == "PLUGINBEGIN") {
-	return plugin_begin(msg);
-    }
-    if (msg.command == "PLUGINEND") {
-	return plugin_end(msg);
-    }
-    if (msg.command == "COMMITBEGIN") {
-	return commit_begin(msg);
-    }
-    if (msg.command == "COMMITEND") {
-	return commit_end(msg);
-    }
-
-    return ZyppPlugin::dispatch(msg);
-}
 
 class SnapperZyppPlugin : public ZyppCommitPlugin {
 public:
